Deleted YumemiWeather from LayoutPrototype.
